Implemented serch via embeddings
