--- Embeddings_Search/Report.md
@@ -1,15 +1,0 @@
-# Что делал
-
-Очень много времени потратил на попытки оптимизировать сканер пдф, но безуспешно
-
-    В итоге распараллелил (на отладку тоже немало времени убил)
-
-Пдф-ки сканируются нормально. Их содержимое LLM воспринимает => на вопросы по содержимому отвечает корректно
-
-Есть проблема с обработкой и восприятием схем - в текущий момент они порождают лишь мусор на выходе сканера
-
-# Дальнейшие мысли
-
-Дальше планировал реализовать разбиение текстов на куски (сейчас на вход LLM подаётся статья целиком)
-
-А так же поиск релевантных текстов через их эмбеддинги
