+++ Embeddings_Search/datasheets.json
@@ -1,0 +1,46 @@
+{
+    "datasheet_searcher_example": {
+        "credentials": "SoMeCreds",
+        "prompt_template": "SoMePrOmPt {context}",
+        "pdfs_location": "",
+        "database_location": ""
+    },
+    "datasheet_searcher": {
+        "chat_model" : "GigaChat",
+        "credentials": "ZjkyMWJhMmUtMTA3OC00NzBmLTk1NTctZjkwZTFlM2EwMjEzOjBkZmNkMzcxLWM3Y2EtNGRmOS1hYzlmLWZjMjBmNGNlMDJlMg==",
+        "prompt_template": "Ты эксперт по поиску информации в тексте. Твоя задача кратко, но в то же время полно ответить на вопрос по тексту.\nОтвечай на языке запроса! При необходимости переведи текст.\nОбязательно УКАЖИ ИСТОЧНИК ИНФОРМАЦИИ - filename!!!\nДалее приведен текст \"{context}\"",
+        "database_location": "C:\\Files\\Science\\3. Использование LLM при проектировании ЭРА\\Embeddings_Search\\data.csv",
+        "pdfs_location": "C:\\Files\\Science\\3. Использование LLM при проектировании ЭРА\\Papers\\"
+    },
+    "logging": {
+        "version": 1,
+        "disable_existing_loggers": true,
+        "formatters": {
+            "standard": {
+                "format": "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
+            },
+            "minimal": {
+                "format": "[%(levelname)s]: %(message)s"
+            }
+        },
+        "handlers": {
+            "default": {
+                "level": "INFO",
+                "formatter": "minimal",
+                "class": "logging.StreamHandler",
+                "stream": "ext://sys.stdout"
+            },
+            "file" : {
+                "level": "DEBUG",
+                "formatter": "standard",
+                "class": "logging.FileHandler",
+                "filename": "mySearcher.log",
+                "encoding": "utf8"
+            }
+        },
+        "root": {
+            "handlers": [ "file" ],
+            "level": "DEBUG"
+        }
+    }
+}
